backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-socketio==5.3.6
flask-cors==4.0.1
psycopg2-binary==2.9.9
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1
python-dotenv==1.0.1
eventlet==0.35.2
gunicorns
